Move source information to a side table
Prior to this change, source information collected for tooling
was as a slot in the group associated with a call. This had two
down-sides, 1) it grew the size of the slot table unnecessarily
and 2) it introduced groups just to record the source information
for functions that otherwise would not need a group.

With this change, the collection of source information is
1) now disabled by default and only enabled when the tooling
API requests it, and 2) moved to a side table instead of being
recorded in the main table.

As source information collection was on by default and is now
off, the size of the slot table for builds that do not use R8,
ProGuard or similar tools are smaller.

Test: ./gradlew :composer:r:r:tDUT
Fixes: 254480106

Change-Id: Ic8b75b1dab148d7cb0313131bd5df5b8a209cc6e