Resolve call extension type in Jetpack ICS
With supporting the new VOIP app actions being added, the Jetpack layer
needs to be able to resolve the extension type for the call before
initiating capability exchange. If the v2 APIs are being used (U/V) or
the backwards compat library is being leveraged (on pre-U devices), we
need to ensure that CAPABILITY_EXCHANGE is the resolved extension type.
For apps that are using the v1.5 Connection Service (AUTO), we need to
resolve the extension type as EXTRAS when the call extras contains
EXTRA_VOIP_API_VERSION. If none of these conditions apply, the phone
account will be checked to see if it supports transactional ops. If the
caller does not have the required permissions, the extension type is
resolved as UNKOWN (no capability exchange). If the phone account
doesn't have the required capability, we can simply resolve the
extension as not supporting capability exchange (NONE).

Fixes: 297438037
Test: atest InCallServiceCompat (on pre-U, U, V devices)
Change-Id: I55aa958e75c28029747a52cc5a8537761a0de430