Provide the context as a preview method parameter instead of TilePreviewData callback parameters
The context is always the same from one tile rendering to another.
If users have classes that depend on the context, having to recreate the object
in the TilePreviewData#onTileRequest can be a bit awkward. Having the context provided
as a preview method parameter makes it easier for the user.

The change also assumes the preview method might be of the wrong type.
This is to prevent an exception from being raised.

Relnote: Provide the context as a preview method parameter instead of TilePreviewData callback parameters
Bug: 227474520
Bug: 298195104
Test: added test cases in TileServiceViewAdapterTest
Change-Id: I5e97d259efc03dd36bb6855288ebfc75bc21d0d7