Add isFromJava() and isFromKotlin() to XMemberContainer.
This CL adds isFromJava() and isFromKotlin() to XMemberContainer to
determine if the container is originally defined in Java or Kotlin
sources or classes.

Note that it's possible for both checks to return false if the container
is an XTypeElement and has an error type.

This CL also adds XProcessingEnv.getElementsFromPackage(). This is
mainly used to get Kotlin top-level functions or properties in KSP. On
the Javac side this is equivalent to getTypeElementsFromPackage() as
Kotlin top-level elements are always wrapped in a type in Javac.

Test: XElementTest
Change-Id: If2bac78591c86d1bb777bbc70d0dcc02c94184b0