Invalidate cache for sensor orientation
When an app isn't subject to foldable-related compatibilty treatment,
the sensor orientation could change depending on the device state. Here
we check the presence of INFO_DEVICE_STATE_SENSOR_ORIENTATION_MAP in
order to determine whether to add sensor orientation to the blocklist so
that we requery it every time.

Bug: 297211960
Test: Build apps on CameraPipe and verify that sensor orientation is
blocklisted.

Change-Id: I35bc3bf56ab2283bf704d7d9e9e6b8cc23b750f6