Check for allowlisted package with single signature
While returning the origin through the getOrigin API, we get the
signature of the calling package and check if it is in the allowlist
passed in by the provider. With this change, we will only
consider a single signature (the most recent one) that the package is
signed with instead of checking against all previous signatures too.

For apps that are signed with multiple signers, all signatures must be
present in the allowlist.

Test: Unit tests. Will follow up with more tests
Bug: 299336848

Change-Id: Ie6ff59011999077b19dafae607bb26f54ce89290