M3 ModalBottomSheet swipe to dismiss fix.
In the M3 AnchoredDraggable migration of ModalBottomSheet one custom implementation of draggable was removed in favor of anchoredDraggable. This generated a bug where swipe to dismiss actions were not calling onDismissRequest after hiding the sheet.

Test: Added tests to verify the new behavior.
Fixes: 298648794
Relnote: Fixed a bug where ModalBottomSheet was not calling onDismissedRequest when dismissing it by swiping down on the sheet.
Change-Id: Idfdd8b490caed6486292ae244ba9e4e9fb813f96