[Overlay] Add an overlay effect API
The OverlayEffect is a base layer API for simple effects. We will be building higher level API such as SyncedOverlayEffect and/or BitmapOverlayEffect on top of it and provide the code samples there.

We don't expect developers to use directly. That is why we didn't provide code samples in the Javadoc because a proper code sample would be too complicated. Still, we want to make this base layer API public to provide maximum flexibility to those who need it.

API guideline conformance changes:
- Rename timestampNs to timestampNanos
- Change OverlayEffect#release to OverlayEffect#close.

Design doc: go/camerax-overlay-effect

Relnote: A OverlayEffect enables the overlaying of graphics on camera frames using OpenGL. It offers a queuing mechanism that allows deferred rendering, ideal for integrating analytical results like QR code highlighting.
Bug: 296461675
Test: manual test and ./gradlew bOS
Change-Id: Ifb7cafd7c88b9a3327a33f5e63b68b0014bcf931