Fix E2ECallExtensionExtrasTests assertion
In Pixel 2 emulators on APIs 30/33, the call extras in the details
arriving in from Telecom were empty bundles, which resulted in test
failures only in the outgoing call scenario. Extending the wait to
include a non-empty entry will resolve the issues seen on older
emulators. Additionally, wrapping the test code in a try/finally block
to ensure that the call is disconnected regardless of any exceptions
thrown will prevent subsequent tests from failing.

Bug: 299949335
Test: androidx.core.core-telecom.androidTest
Change-Id: I2748f6314c182475adea2b0628a2b71d9b0e9762