Set transition state after animateToStart
After a call to animateToStart in order to get transitions back to the
original state you need to make a call to beginDelayedTransition.
Failing to do so will not clear the view overlay and cause subsequent
transitions to fail.

RelNote: "Fixed an issue where the screen would go blank during a
gesture back after the first time cancelling the transition."
Test: coming

Change-Id: I02f8be17759ef7c90baded79c06c5674405a3665