Resolve pre-U audio suppression issues.
In the legacy Jetpack path, audio suppression was being observed as a
result of the Jetpack connection not explicitly setting voip audio mode.
Once the connection is successfully created and Telecom is notified, the
VOIP mode of the call is updated to false, which leads to audio
suppression.

Through modifying the integration test app to observe AudioRecords, I
was able to verify that the calls are not being suppressed (even when
the app is backgrounded for over a minute).

Bug: 291955208
Test: integration test app
Change-Id: If050f5e68d0f1a73e61941fecf49c0d3c96a400e