Adds IndicationNodeFactory and deprecates Indication#rememberUpdatedInstance
The existing @Composable factory function on Indication results in a high-cost performance floor: it requires us to use Modifier.composed, and
prevents us from doing any lazy initialization of indication. This cl adds a new IndicationNodeFactory, which behaves similarly to
ModifierNodeElement in terms of responsibilities and semantics, which allows us to handle the stateful requirements of indication nodes by
relying on the existing Modifier node infrastructure, instead of modelling that in a composable function. This should lead to simpler
Indication implementations that have more in common with draw modifiers, as well as improved performance and room for future performance
optimizations in clickable and other components.

This CL suppresses the deprecation on existing Indication implementations, they will be rewritten in a future CL.

Bug: b/298048146
Test: updateApi
Relnote: "Deprecates Indication#rememberUpdatedInstance, and adds IndicationNodeFactory as a replacement. See developer.android.com for migration information and more background information behind this change."
Change-Id: I63537be74c7090aec9895764c13bfdfb9e806a93