Fix CaptureSession configure fail on legacy devices
The solution is to wait for all in-flight captures to be processed
before starting a new CameraCaptureSession.

(1) Create IncorrectCaptureStateQuirk for this error.
(2) Rename WaitForRepeatingRequestStart to RequestMonitor, which is
 a more descriptive name.
(3) Update RequestMonitor so that getRequestStartedFuture can track both
 repeating and single capture sequences.
(4) Only start creating the new CameraCaptureSession after all the in-flight
 capture sequences are processed.
(5) Set a 1.5-second timeout for the in-flight captures.

Bug: 297501750
Test: CaptureSessionTest & Camera2CameraImplTest & CameraControllerFragmentTest
Change-Id: Ic843985739a4aa0a5775db178e77f0c70bc87e80