Fix a datepicker crash when navigating months fast
Added a try-catch guard to prevent a crash when clicking the navigation
arrows fast and reaching the end of the provided year range.
Opted for a try-catch instead of checking the LasyListState
"isScrollInProgress" for the buttons enablement. Using the in-progress
will disable the buttons until the transition settles, which makes it
hard to navigate by quickly tapping the buttons.

Bug: 290954897
Bug: 297002119
Bug: 290954897
Test: Manual
Relnote: "Fix a DatePicker crash when quickly clicking the navigation
arrow buttons when the displayed month is at the edge of the allowed
range of years."

Change-Id: I46f360883dc57d677a545863ff2d913903861928