Add GlanceAppWidget.runComposition
This change adds GlanceAppWidget.runComposition which returns a
Flow<RemoteViews>. This API can be used to run the widget outside of the
app widget system (e.g. without a GlanceAppWidgetReceiver or needing to
be bound by the AppWidgetHost) and without a Worker. The widget will
continue to recompose and respond to lambda actions while the flow is
running.

This makes it straightforward to display a widget within an Activity,
and is intended to be used for the glance-experimental-tools
WidgetViewer, which provides Live Edit in Studio.

Bug: 298066147
Test: SimpleWidgetViewer demo activity, GlanceAppWidgetTest unit tests
Relnote: "Add GlanceAppWidget.runComposition"
Change-Id: I6344ce93a486a86b7484dfd704453da118e3aa79