Enable SandboxedSdkViewTest on older SDK version
Test related to input token fails on API 30. This is due to how we have
implemented removeSurfaceViewAndOpenSession() in SandboxedSdkView.kt
file. On API level lower than 33, we simply pass a placeholder IBinder
token. Our test doesn't take placeholder token in consideration, so it
fails.

Input token is needed for creating SCVH and SCVH is needed in UI Lib
only on devices that support SdkSandbox on a separate process. So it
should be safe to ignore the test on devices where provide cannot be
remote.

Created a separate BackwardCompat utility class to highlight forks in
logic instead of sprinkling build verion checks in code base.

Bug: 300397160
Bug: 298618670
Test: presubmit
Change-Id: Id11a2d33bfe377d33f699b3f14b48661c5908f8d