Fix content node in scrolling ANI tests
Previously testSortedAccessibilityNodeInfo_ScaffoldScrollingTopBar had a testTag set on a column instead of the actual text content in a testing scene. The test assertions still passed because the column generated a traversal value of 0; since traversal values should be positive numbers, this change adds in an assertion to make sure the traversal values aren't 0 in addition to putting the testTag in the correct location.

Test: testSortedAccessibilityNodeInfo_ScaffoldScrollingTopBar()
Change-Id: I9a7cc209e98c44dae54a6a1088dcce0b8322aa0b