Let GattServer know the subscribed characteristics
This change adds a state flow used to track the
set of characteristics that the client has requested
to be notified of.

It also cleans up how to deal with notification and
indication: prefer notification over indicaiton
under the hood.

Bug: 303768910
Test: ./gradlew bluetooth:bluetooth:check
bluetooth:bluetooth:connectedCheck
bluetooth:bluetooth-testing:check
Relnote: Added GattServerSessionScope#subscribedCharacteristics

Change-Id: I0edabd53e1af3dcd64abd42470887a36de236144