Start setting lint version for androidx-studio-integration
In buildSrc we add several libraries such as tools common that are
loaded by the version defined in libs.versions.toml. This means that
we end up loading an older version of these libraries instead of using
just built version of AGP + tools common. This fixes the current
integration build failure where we try to load aapt2 version
aapt2-8.3.0-alpha04-10880808 instead of aapt2-8.3.0-dev-10880808
because we use AGP's gradle-8.3.0-dev, but common is 8.3.0-alpha04
as defined in libs.versions.toml

Test: ./busytown/androidx-studio-integration.sh -> no longer failing
      in configuration phase.
Change-Id: I5c88efe163e68bdcf3db8018b6c50f0c1de50ced