Reduce calls to CompositionLocalProvider
Providing a new value to a composition local provider is expensive
and cost can be saved by providing multiple values in one provider call.
This CL introduces a convenience method to provide both a content color
and text style in the same call as this is a common pattern in the code base.

Fixes: 297211038
Test: Existing
Relnote: N/A
Change-Id: I91379372ed179ad3ee25d1d14ff93a6df70d4881