Make PickVisualMediaRequest API 19 to match contract
The PickVisualMedia Activity Result contract
is labeled as RequiresApi(19). To launch that
contract, you need to create a PickVisualMediaRequest,
but that PickVisualMediaRequest object was not
labeled as RequiresApi(19) and was using constants
from PickVisualMedia.

Both of these APIs should match and both use
RequiresApi(19).

Test: ./gradlew activity:activity:lintDebug
Relnote: "`PickVisualMediaRequest` now has the same
minimum API level of 19 as the `PickVisualMedia`
Activity Result contract."

Change-Id: Id6e21cd21f91dd81bcb1c3b23286e5528ffa7cd6