Rename Foundation/SwipeToDismissBox to Foundation/BasicSwipeToDismissBox
We have renamed the Foundation level SwipeToDismissBox to BasicSwipeToDismissBox. This makes the distinction clearer between the Foundation level component and the Material level SwipeToDismissBox. The latter pulls colors from the MaterialTheme to be used in scrims and delegates the remaining implementation to the BasicSwipeToDismissBox.

Test: androidx.wear.compose.integration.macrobenchmark.test
Bug: 292537483
RelNote: "We have renamed the Foundation level SwipeToDismissBox to BasicSwipeToDismissBox. This makes the distinction clearer between the Foundation level component and the Material level SwipeToDismissBox. The latter pulls colors from the MaterialTheme to be used in scrims and delegates the remaining implementation to the BasicSwipeToDismissBox."
Change-Id: Ibecfc1720eadd2d8e5e1c1cfd6832300775bffb1