Additions to SeekableTransitionState
Fixes: 301414577

Relnote: "SeekableTransitionState has been updated
to enable transitioning between multiple states
instead of two states. Seeking can be done with
seekTo() and snapToFraction(). Animation can be
done with animateTo() or animateToTargetState.

When a transition is interrupted, the initial
values will start or continue to animate to the
former targetState values. Only control over the
currentState to targetState are given to the
developer."

Test: new tests, manual tests

Change-Id: I555fd7be787c1962f1692882254ed8dca41d1579