Simple changes to fix issue b/204328129
Please see https://issuetracker.google.com/issues/204328129 for more details on the issue.
Basically this is something necessary to help prevent weird compilation errors that happen with the Buck/Redex build toolchains that Meta uses. Not something meaningful, but since the change is dead simple I guess it shall pass.

Test: This has been something that was patched on Meta source and we've been testing it for multiple prod releases of Threads, we're only now porting the change to Google because it seems appropriate.
Fixes: b/204328129