Move to using custom anchor task for running KMP tests
allTests anchor task from Kotlin Gradle Plugin has issues as it
tries to control the ignoreFailures and upToDateIf properties of
test tasks when it is added to the task graph. Instead, use allHostTests
anchor task that does not do this. This task was added in aosp/2799456

This change means than when a test fails, it will no longer break the build
the way it is doing now for datastore test flakes.

Change-Id: If5501e88891e44b6ddaf3c4a9c1c068ac354f95d
Test: None