Disable system prediction by default
At the moment there are no heuristics applied to the system
prediction, which makes the errors noticeable on some implementations.

This change will prefer the internal implementation for the time being;
a new property has been added, `debug.input.prefer_system_prediction`,
which will reenable the system prediction if present.

Test: `adb shell setprop debug.input.prefer_system_prediction true`
Bug: 302300930
Change-Id: Ic36bcb894d6bce319882f456840bcd2f9d67822b