Fix mac build
Opt-in to cinterop APIs that became experimental in 1.9.20
(ExperimentalForeignApi)

Migrate to use CPointer instead of underlying type in SQLite native.
It avoids the .pointed and .ptr back-and-forth and is compatible with Kotlin 1.9.0 changes.

Remove KMP mem allocator override
This was only necessary for overiding mimalloc since it wasn't supported
on Android, but we can now rely on the default custom memory allocator
in 1.9.20 instead, which will also have better performance.

Added a workaround for metadata generation in the benchmark plugin:
KT-42387

Bug: 308798582
Test: NativeSQLiteDriverTest
Test: CI
Change-Id: I500bfcce6fad7a36761fc22fd88aef2b8f223868