[Tooltip] Adding caret to Plain Tooltip APIs
Adds the capability to enable carets to show on plain tooltips via PlainTooltipBox and PlainTooltip. Added a new CaretProperties class that contains the caret's dimensions. Adding CaretScope interface which will provide the tooltip parameter with the anchor's bounds which can be used to calculate where the caret should point to.

Test: Adding unit tests for the positioning logic of plain and rich tooltips when they would collide with the top of the screen. Adding caret-anchor positioning unit test to ensure that the anchor bounds provided to the DrawScope of the caret is correct.
Bug: 297037973
Relnote: Adding the capability to enable tooltip carets for Plain Tooltip APIs.
Change-Id: Ibf767f249315a4c4c327b62a487c6770435c13c7