Update Checkbox, Switch and RadioButton to meet accessibility standards
These components do not meet minimum tap target sizes (48dp) - given that they are intended for use only in ToggleButton as a status display, we will remove the click handling.

Bug: 297341220
Test: Updated SelectionControlTests
Relnote: "We have changed the Checkbox, Switch and RadioButton components to be display-only, by removing the click handling. These components are expected to be used in (Split)ToggleButton which handles the click, so the components are now more clearly indicated as display-only (and are not intended for standalone use on Wear)."
Change-Id: I2322ed56b42d364dfaa98e7ba72ff62b4fb5b9c3