Support hyperlinks in text
This CL adds basic functionality to support links without introducig new APIs.
Use UrlAnnotation to add a link and by default it will open it using the UriHandler. Only taps are supported in this CL, no other gestures.

Internally, we add focusable invisible layout nodes on top of each text range with the UrlAnnotation. We also support key events and add clipping to the path (path represents the link's range).

In the upcoming CLs we'll add
- API for styling
- new annotations to support a general clickable in addition to a link
- accessibility
- click handler API
- other gestures support

Test: BasicTextLinkTest

Change-Id: I78401712be7e70d9d149a9480c7eec7fcf21aa35