Fix flakes in testAssetLoaderAjaxActivity
The flakes were caused by a race between JavaScript execution and
loading of the HTML elements, leading to the elements not being
available by the time the JavaScript tries to access it.

Resolved by reordering the html asset to put the output HTML elements
before the JavaScript.

Also adds a button to the test app to manually load the HTML when not
used by the instrumentation test.

Test: AssetLoaderAjaxActivityTestAppTest.java
Fixes: 267242191
Change-Id: Ib9ce3c65282b6e0e9e708037b3777a16ce1ebdbd