Attempt to fix flaky focus/keyboard tests for BTF and BTF2
Instead of testing the visibility of keyboard, it should be sufficient to just test whether a text input session is started. Showing the keyboard is the eventual outcome of starting an input session. We do not have to test this explicitly because waiting for keyboard can timeout for numerous reasons.

Fix: b/310278572
Test: compose:found:found:cAT
Change-Id: I58e2f65ab4c136f21004ba0466ef6861e4c99adc