Adds ripple API to wear material and material3 libraries
The new ripple API replaces the deprecated rememberRipple API - it creates a ripple instance that will use the default theme values. ripple
does not need to be remembered, similar to modifiers - if the parameters compare equally, then the same node can be reused internally.

This CL also migrates all wear:compose-material / wear:compose-material3 components to use the new ripple. There is a temporary
CompositionLocal, LocalUseFallbackRippleImplementation, which can be set to true to fallback to using the old RippleTheme / rememberRipple
implementation, but this is strongly discouraged as it is much less performant than the new implementation. This fallback will be removed in
the next stable release, and exists here to aid migration.

Bug: b/298048146
Bug: b/304985887
Test: RippleTest
Relnote: "Adds new ripple API in wear:compose-material and wear:compose-material3 libraries which replaces the deprecated rememberRipple. Also
adds a temporary CompositionLocal, LocalUseFallbackRippleImplementation, to revert Material components to using the deprecated rememberRipple
/ RippleTheme APIs. This will be removed in the next stable release, and is only intended to be a temporary migration aid for cases where you
are providing a custom RippleTheme. See developer.android.com for migration information and more background information behind this change."

Change-Id: I87fd7d3968ec14c3e684039298b31a1fc620b47c